+++ tests/testthat/test_model_weighting.R
@@ -1,7 +1,7 @@
 library(loo)
 options(loo.cores = 2)
 
-context("model_weights")
+context("loo_model_weights")
 
 # generate fake data
 set.seed(123)
@@ -21,23 +21,23 @@ r_eff_list <- list(rep(0.9,50), rep(0.9,50), rep(0.9,50))
 
 tol <- 0.01 # absoulte tolerance of weights
 
-test_that("model_weights throws correct errors", {
-  expect_error(model_weights(log_lik1), "is.list")
-  expect_error(model_weights(list(log_lik1)), "At least two models")
-  expect_error(model_weights(list(log_lik1, log_lik2[-1, ])), "same dimensions")
-  expect_error(model_weights(list(log_lik1, log_lik2, log_lik3[, -1])), "same dimensions")
+test_that("loo_model_weights throws correct errors", {
+  expect_error(loo_model_weights(log_lik1), "is.list")
+  expect_error(loo_model_weights(list(log_lik1)), "At least two models")
+  expect_error(loo_model_weights(list(log_lik1, log_lik2[-1, ])), "same dimensions")
+  expect_error(loo_model_weights(list(log_lik1, log_lik2, log_lik3[, -1])), "same dimensions")
 
-  expect_error(model_weights(ll_list, r_eff_list = r_eff_list[-1]),
+  expect_error(loo_model_weights(ll_list, r_eff_list = r_eff_list[-1]),
                "one component for each model")
 
   r_eff_list[[3]] <- rep(0.9, 51)
-  expect_error(model_weights(ll_list, r_eff_list = r_eff_list),
+  expect_error(loo_model_weights(ll_list, r_eff_list = r_eff_list),
                "same length as the number of columns")
 })
 
 
-test_that("model_weights (stacking and pseudo-BMA) gives expected result", {
-  w1 <- model_weights(ll_list, method = "stacking", r_eff_list = r_eff_list)
+test_that("loo_model_weights (stacking and pseudo-BMA) gives expected result", {
+  w1 <- loo_model_weights(ll_list, method = "stacking", r_eff_list = r_eff_list)
   expect_type(w1,"double")
   expect_s3_class(w1, "stacking_weights")
   expect_length(w1, 3)
@@ -45,7 +45,7 @@ test_that("model_weights (stacking and pseudo-BMA) gives expected result", {
   expect_equal_to_reference(as.numeric(w1), "model_weights_stacking.rds",
                             tolerance  = tol, scale=1)
 
-  w2 <- model_weights(ll_list, r_eff_list=r_eff_list,
+  w2 <- loo_model_weights(ll_list, r_eff_list=r_eff_list,
                       method = "pseudobma", BB = TRUE)
   expect_type(w2, "double")
   expect_s3_class(w2, "pseudobma_bb_weights")
@@ -54,7 +54,7 @@ test_that("model_weights (stacking and pseudo-BMA) gives expected result", {
   expect_equal_to_reference(as.numeric(w2), "model_weights_pseudobma.rds",
                             tolerance  = tol, scale=1)
 
-  w3 <- model_weights(ll_list, r_eff_list=r_eff_list,
+  w3 <- loo_model_weights(ll_list, r_eff_list=r_eff_list,
                       method = "pseudobma", BB = FALSE)
   expect_type(w3,"double")
   expect_length(w3, 3)